+++ src/egon/data/datasets/electrical_neighbours.py
@@ -12,6 +12,7 @@ import egon.data.datasets.etrago_setup as etrago
 import egon.data.datasets.scenario_parameters.parameters as scenario_parameters
 from egon.data import config, db
 from egon.data.datasets import Dataset
+from egon.data.datasets.fill_etrago_gen import add_marginal_costs
 from egon.data.datasets.scenario_parameters import get_sector_parameters
 
 
@@ -19,7 +20,7 @@ class ElectricalNeighbours(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="ElectricalNeighbours",
-            version="0.0.6",
+            version="0.0.7",
             dependencies=dependencies,
             tasks=(grid, {tyndp_generation, tyndp_demand}),
         )
@@ -941,15 +942,22 @@ def insert_generators(capacities):
         get_foreign_bus_id().loc[gen.loc[:, "Node/Line"]].values
     )
 
+    # Add scenario column
+    gen["scenario"] = "eGon2035"
+
+    # Add marginal costs
+    gen = add_marginal_costs(gen)
+
     # insert generators data
     session = sessionmaker(bind=db.engine())()
     for i, row in gen.iterrows():
         entry = etrago.EgonPfHvGenerator(
-            scn_name="eGon2035",
+            scn_name=row.scenario,
             generator_id=int(db.next_etrago_id("generator")),
             bus=row.bus,
             carrier=row.carrier,
             p_nom=row.cap_2035,
+            marginal_cost=row.marginal_cost,
         )
 
         session.add(entry)
@@ -1008,10 +1016,7 @@ def insert_generators(capacities):
             series_egon100.generator_id == egon_2035_to_100[gen_id]
         ]["p_max_pu"].values[0]
         entry = etrago.EgonPfHvGeneratorTimeseries(
-            scn_name="eGon2035",
-            generator_id=gen_id,
-            temp_id=1,
-            p_max_pu=serie,
+            scn_name="eGon2035", generator_id=gen_id, temp_id=1, p_max_pu=serie
         )
 
         session.add(entry)
@@ -1129,7 +1134,7 @@ def get_map_buses():
         "SE01": "SE02",
         "SE03": "SE02",
         "SE04": "SE02",
-        "RU":   "RU00",
+        "RU": "RU00",
     }