+++ src/egon/data/datasets/heat_supply/__init__.py
@@ -170,10 +170,15 @@ class HeatSupply(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="HeatSupply",
-            version="0.0.6",
+            version="0.0.7",
             dependencies=dependencies,
             tasks=(
                 create_tables,
-                {district_heating, individual_heating, potential_germany},
+                {
+                    district_heating,
+                    # Temporary drop everything related to rural heat
+                    # individual_heating,
+                    potential_germany,
+                },
             ),
         )