+++ src/egon/data/datasets/hydrogen_etrago/__init__.py
@@ -67,7 +67,7 @@ class HydrogenMethaneLinkEtrago(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="HydrogenMethaneLinkEtrago",
-            version="0.0.4",
+            version="0.0.5",
             dependencies=dependencies,
             tasks=(insert_h2_to_ch4_to_h2, insert_h2_to_ch4_eGon100RE),
         )